+++ cli/src/commands/git/remote/add.rs
@@ -21,7 +21,6 @@ use jj_lib::repo::Repo;
 
 use crate::cli_util::CommandHelper;
 use crate::command_error::CommandError;
-use crate::commands::git::FetchTagsMode;
 use crate::git_util::absolute_git_url;
 use crate::ui::Ui;
 
@@ -36,10 +35,6 @@ pub struct GitRemoteAddArgs {
     /// Local path will be resolved to absolute form.
     #[arg(value_hint = clap::ValueHint::Url)]
     url: String,
-
-    /// Configure when to fetch tags
-    #[arg(long, value_enum, default_value_t = FetchTagsMode::Included)]
-    fetch_tags: FetchTagsMode,
 
     /// The URL used for push
     ///
@@ -63,13 +58,7 @@ pub async fn cmd_git_remote_add(
 
     let mut tx = workspace_command.start_transaction();
 
-    git::add_remote(
-        tx.repo_mut(),
-        &args.remote,
-        &url,
-        push_url.as_deref(),
-        args.fetch_tags.as_fetch_tags(),
-    )?;
+    git::add_remote(tx.repo_mut(), &args.remote, &url, push_url.as_deref())?;
     warn_if_remote_url_matches(ui, tx.repo(), &args.remote, &url, push_url.as_deref())?;
     tx.finish(ui, format!("add git remote {}", args.remote.as_symbol()))
         .await?;